style: apply ruff format across hamon, tests, and benchmarks
Formatting-only; brings the 33 files that predated the formatter into
compliance so ruff format --check passes repo-wide.

+++ hamon/conditional_samplers.py
@@ -193,6 +193,4 @@ class SoftmaxConditional(AbstractParametricConditionalSampler):
         output_sd: PyTree[jax.ShapeDtypeStruct],
     ) -> tuple[_State, None]:
         """Sample from a softmax distribution given the parameter vector $\theta$."""
-        return jax.random.categorical(key, parameters, axis=-1).astype(
-            output_sd.dtype
-        ), sampler_state
+        return jax.random.categorical(key, parameters, axis=-1).astype(output_sd.dtype), sampler_state
